+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
+## 2016.4.11
+
+* change isSingle option to 'single'
+* add process percent / size / filename css
+* add more demo
+* release 1.0
+
 ## 1.0.2
 
 * OMD: this version is written on omd 1.1.2, look at https://github.com/tangshuang/omd